Move buildPreset to the separate module.
